+++ lib/handlers.dart
@@ -5,3 +5,5 @@ export './src/handlers/blank.dart';
 export './src/handlers/direct.dart' show direct;
 export './src/handlers/parser.dart' show HandlerType;
 export './src/handlers/json.dart' show JsonCTX, json;
+export './src/handlers/form.dart' show FormCTX, form;
+export './src/handlers/iform.dart' show iFormCTX, iForm;